+++ CHANGELOG.md
@@ -13,6 +13,8 @@ All notable changes to this project will be documented in this file.
     * this allows users and compare and hash `BmpPerPeerHeader` structs
     * also implemented `.strip_timestamp()` to remove the timestamp from the `BmpPerPeerHeader` struct for cases where
       the timestamp is not needed
+* rename `MessageBody` to `BmpMessageBody`
+* derive `Clone`, `PartialEq` to `BmpMessage` and `MessageBody`
 
 ## v0.10.8 -2024-04-05