+++ CHANGELOG.md
@@ -15,6 +15,7 @@ All notable changes to this project will be documented in this file.
       the timestamp is not needed
 * rename `MessageBody` to `BmpMessageBody`
 * derive `Clone`, `PartialEq` to `BmpMessage` and `MessageBody`
+* added serialization/deserialization support for `BmpMessage` and `BmpMessageBody`
 
 ## v0.10.8 -2024-04-05